230111 김다희 게시판 끝

+++ BookingCatch/src/main/java/com/smhrd/model/tipBoardVO.java
@@ -26,7 +26,7 @@ public class tipBoardVO {
 	@NonNull
 	private String writer;
 	@NonNull
-	private int likes;
+	private BigDecimal likes;
